requirements.txt
requests>=2.28.1
pyyaml>=6.0
jsonschema>=4.17.3
python-dateutil>=2.8.2
pyjwt>=2.6.0
sqlalchemy>=2.0.0
colorama>=0.4.6
rich>=13.0.0
aiohttp>=3.8.4
pytest>=7.3.1
pytest-asyncio>=0.21.0
cryptography>=40.0.2
redis>=4.5.4
ratelimit>=2.2.1

# Optional dependencies for enhanced JWT scanning
# PyJWT is already included above (pyjwt>=2.6.0)
jwcrypto>=1.4.2